Change entity fields from class to instance attrs
Representing entity fields as class attributes works well if the underlying
schema you are targeting is unchanging. However, Satellite's schema changes over
time. For example, one difference between Satellite 6.0 and Satellite 6.1 is
that the latter added a "gpg_key" field to the Product entity. This is a problem
if you are working with two different servers, one running 6.0 and the other
running 6.1.

Representing entity fields as instance attributes allows the exact set of fields
on a given entity to be determined at run-time. Thus, you could have one Product
object that targets 6.0 and does not have a gpg_key field and a second Product
object that targets 6.1 and does have a gpg_key field.

This commit targets, but does not fix, #15:

> Make it possible to interact with multiple versions of the Satellite API using
> NailGun.

Using instance attributes produces more verbose code. However, it also allows
some nasty hacks to be dropped:

> There is at least one *very* nice side-effect of this approach. Currently, the
> entities contain lines such as the following:
>
>     gpg_key = entity_fields.OneToOneField('GPGKey')
>
> Using instance attributes allows us to name classes directly:
>
>     'gpg_key': entity_fields.OneToOneField(GPGKey)
>
> Naming classes directly would allow us to drop fragile stuff like
> `nailgun.entity_fields.ENTITIES_MODULE` and
> `nailgun.entity_fields._get_class`.